requirements.txt
docker-py==1.7.2
jinja2